fix TagCloud

+++ app/models/tag_cloud.rb
@@ -1,17 +1,31 @@
 class TagCloud < ActiveRecord::Base
-
+  # TagCloud.generate
   def self.generate
     Lastsearch.delete_old_rows
-    rezs = Lastsearch.find(:all, :order => "num DESC", :limit => 21) # Допустим num = 20..100
-    max = rezs.first.num                                         # max = 100
-    min = rezs.last.num                                         # min = 20
-    rang = (max - min) / 15                                      # 15, помежуток в пикселях между минимальним и максимальным шрифтом (100 - 20) делим на  15, получаем шаг одной позиции пикселя равный 5.3 единиц
-    TagCloud.delete_all # Очищаем таблину от текущих тегов.
-    for rez in rezs
-      font_size = ((rez.num - min) / rang).to_i + 10             # Вычисляем размер шрифта 100 - 20 делим на 5.3 = 15 + 10 минимальный размер шрифта, и получаем шрифт равный 25px
-      TagCloud.create(:font_size => font_size, :url_string => rez.url_string, :url_attributes => rez.url_attributes, :url_model => rez.url_model)
+    @last_searhes = Lastsearch.all(:select =>"count(*) as count_items, url_string, url_attributes, url_model",
+                                   :group => "url_string, url_attributes, url_model",
+                                   :order => "count_items DESC",
+                                   :limit => 21)
+    if @last_searhes # Допустим num = 20..100
+      @max_num = @last_searhes.first["count_items"].to_i
+      @min_num = @last_searhes.last["count_items"].to_i
+
+      TagCloud.delete_all # Очищаем таблину от текущих тегов.
+      @last_searhes.each do |item|
+
+        font_size = (( item["count_items"].to_i / @max_num.to_f) * (4 - 1)).round # Взято из плагина https://github.com/mbleigh/acts-as-taggable-on
+
+        TagCloud.create(:font_size      => font_size,
+                        :url_string     => item["url_string"],
+                        :url_attributes => item["url_attributes"],
+                        :url_model      => item["url_model"])
+
+
+      end
     end
+
   end
 
 end
+
